vnurepo/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blessings==1.6.1
bpython==0.17
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
curtsies==0.2.11
dateparser==0.6.0
enum34==1.1.6
greenlet==0.4.12
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
lxml==4.1.1
mysql-connector==2.1.4
parsel==1.3.1
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.3
queuelib==1.4.2
regex==2018.1.10
requests==2.18.4
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.35
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
tzlocal==1.5.1
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
zope.interface==4.4.3
